+++ packages/terra-clinical-application/tests/wdio/application-spec.js
@@ -1,29 +1,23 @@
-Terra.viewports('tiny', 'small', 'medium', 'large', 'huge', 'enormous').forEach((viewport) => {
-  describe('Clinical Application', () => {
+Terra.describeViewports('Clinical Application', ['tiny', 'enormous'], () => {
+  describe('Renders the Application with provided AppDelegate', () => {
     before(() => {
-      browser.setViewportSize(viewport);
+      browser.url('/#/raw/tests/terra-clinical-application/clinical-application/default-application');
     });
 
-    describe('Renders the Application with provided AppDelegate', () => {
-      before(() => {
-        browser.url('/#/raw/tests/terra-clinical-application/clinical-application/default-application');
-      });
+    it('has the app delegate', () => {
+      const pageMessage = browser.element('#test-ContainerComponent').getText();
+      expect(pageMessage).to.include('is present');
+    });
+  });
 
-      it('has the app delegate', () => {
-        const pageMessage = browser.element('#test-ContainerComponent').getText();
-        expect(pageMessage).to.include('is present');
-      });
+  describe('Renders the Application without provided AppDelegate', () => {
+    before(() => {
+      browser.url('/#/raw/tests/terra-clinical-application/clinical-application/no-app-delegate-application');
     });
 
-    describe('Renders the Application without provided AppDelegate', () => {
-      before(() => {
-        browser.url('/#/raw/tests/terra-clinical-application/clinical-application/no-app-delegate-application');
-      });
-
-      it('does not have the app delegate', () => {
-        const pageMessage = browser.getText('#test-ContainerComponent');
-        expect(pageMessage).to.include('is not present');
-      });
+    it('does not have the app delegate', () => {
+      const pageMessage = browser.getText('#test-ContainerComponent');
+      expect(pageMessage).to.include('is not present');
     });
   });
 });