+++ packages/terra-clinical-detail-view/tests/wdio/DetailView/detail-view-spec.js
@@ -1,39 +1,33 @@
 const selector = '#root';
 
-Terra.viewports('tiny', 'small', 'medium', 'large', 'huge').forEach((viewport) => {
-  describe('Detail View', () => {
+Terra.describeViewports('Detail View', ['tiny', 'small', 'medium', 'large', 'huge'], () => {
+  describe('Displays a divided Detail View with provided components', () => {
     before(() => {
-      browser.setViewportSize(viewport);
+      browser.url('/#/raw/tests/terra-clinical-detail-view/clinical-detail-view/detail-view/detail-view-divided');
     });
 
-    describe('Displays a divided Detail View with provided components', () => {
-      before(() => {
-        browser.url('/#/raw/tests/terra-clinical-detail-view/clinical-detail-view/detail-view/detail-view-divided');
-      });
+    Terra.it.validatesElement({ selector });
+  });
 
-      Terra.it.validatesElement({ selector });
+  describe('Displays a Detail View that is not divided with provided components', () => {
+    before(() => {
+      browser.url('/#/raw/tests/terra-clinical-detail-view/clinical-detail-view/detail-view/detail-view-no-divider');
     });
 
-    describe('Displays a Detail View that is not divided with provided components', () => {
-      before(() => {
-        browser.url('/#/raw/tests/terra-clinical-detail-view/clinical-detail-view/detail-view/detail-view-no-divider');
-      });
+    Terra.it.validatesElement({ selector });
+  });
 
-      Terra.it.validatesElement({ selector });
+  describe('Displays a Detail View with provided components and when title size is smaller', () => {
+    before(() => {
+      browser.url('/#/raw/tests/terra-clinical-detail-view/clinical-detail-view/detail-view/detail-view-divided-smaller-titles');
     });
 
-    describe('Displays a Detail View with provided components and when title size is smaller', () => {
-      before(() => {
-        browser.url('/#/raw/tests/terra-clinical-detail-view/clinical-detail-view/detail-view/detail-view-divided-smaller-titles');
-      });
+    Terra.it.validatesElement({ selector });
+  });
 
-      Terra.it.validatesElement({ selector });
-    });
+  describe('Displays a Detail View with overflowed title text', () => {
+    before(() => browser.url('/#/raw/tests/terra-clinical-detail-view/clinical-detail-view/detail-view/detail-view-overflow-text'));
 
-    describe('Displays a Detail View with overflowed title text', () => {
-      before(() => browser.url('/#/raw/tests/terra-clinical-detail-view/clinical-detail-view/detail-view/detail-view-overflow-text'));
-
-      Terra.it.validatesElement({ selector });
-    });
+    Terra.it.validatesElement({ selector });
   });
 });