+++ packages/terra-clinical-item-display/CHANGELOG.md
@@ -3,6 +3,8 @@ ChangeLog
 
 Unreleased
 ----------
+### Changed
+* Added new prop for icon alignment and missing default primary text style
 
 3.3.0 - (March 26, 2019)
 ----------