+++ packages/terra-clinical-item-display/CHANGELOG.md
@@ -3,6 +3,8 @@ ChangeLog
 
 Unreleased
 ----------
+### Removed
+* Removed node_modules from .npmignore
 
 3.3.0 - (March 26, 2019)
 ----------